requirements.txt
langchain>=0.3.0,<1.0
langchain-core>=0.3.0,<1.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
pydantic>=2.0
python-dotenv>=1.0
